I hereby agree to the terms of the [Singularity Data, Inc. Contributor License Agreement](https://gist.github.com/skyzh/0663682a70b0edde7ae991492f2314cb#file-s9y_cla).

## What's changed and what's your intention?

Rewrite `OrderByExecutor` using `memcomparable` and `sort_unstable_by` in `core::slice`.

(I planned to use radix sort but the benchmark result shows that `sort_unstable_by` is fast enough)

## Checklist

- [x] I have written necessary rustdoc comments
- [x] I have added necessary unit tests and integration tests
- [x] All checks passed in `./risedev check` (or alias, `./risedev c`)

## Micro benchmark

*Row count: 1024 * 1024. Columns: (Int64)*
| Batch size | Before | After |
| --- | --- | --- |
| 32 | 2.5112 s |**583.48 ms**|
| 128 |1.6179 s |**480.14 ms** |
| 512 | 1.2122 s|**458.19 ms**|
| 1024 |1.0953 s |**446.31 ms**|
| 2048 |973.76 ms |**445.77 ms**|

*Row count: 1024 * 1024. Columns: (Int64, Varchar, Float32, Timestamp)*
| Batch size | Before | After |
| --- | --- | --- |
| 32 |3.2621 s |**1.3463 s**|
| 128 |2.4378 s |**1.2279 s**|
| 512 |1.8293 s |**1.2024 s**|
| 1024 |1.7531 s |**1.1690 s**|
| 2048 |1.6792 s |**1.1642 s**|

## Refer to a related PR or issue link (optional)
close https://github.com/singularity-data/risingwave/issues/3538